requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9
Pillow>=10.3.0
torch>=2.5.0
diffusers>=0.30.0
transformers>=4.40.0
accelerate>=0.30.0
